v3.1
A Configuration File subsystem has been added. It lets you store and load Bool, Int, Float, FVector, and String values from a file.

For now, it’s optional and only used in a small example. The example is at the start of GUI::Draw() in GUI.cpp and is wrapped in #ifdef _DEBUG.

Switch your project to Debug mode (or remove the #if) to make the new “Config Test” menu appear when you inject UETools-GUI into a title.